[feature] 게임 플레이 전적 더 보기 작업중

+++ opgg/opgg/Sources/Repository/OpggRepository/OpggRepositoryImpl.swift
@@ -15,7 +15,7 @@ final class OpggRepositoryImpl: NetworkRepository<OpggTarget>, OpggRepository {
             .map(SummonerInfo.self)
     }
     
-    func requestMatches(lastMatch: UInt64?) -> Single<Result<Matches, APIError>> {
+    func requestMatches(lastMatch: Int?) -> Single<Result<Matches, APIError>> {
         provider
             .request(.requestMatches(lastMatch: lastMatch))
             .map(Matches.self)
